refactor(storefront/product-card): wrappers spread SURFACE_PRESETS.
CollectionProductCard, RecommendationProductCard, SearchProductCard become
thin spreads over the orchestrator. search-content-gate's skeleton uses
chrome="frameless" (renamed from the dropped "bare"). CartDrawerProductCard
deleted — zero consumers; cart drawer items render via CartLine.
